Home Screen Endpoint Updates
Updated the 'getHomeScreenInfo' method to return total number of homes, locations, and devices for the logged in user. This information will be displayed on the home screen.

Updated 'getDevicesByLocation' method in Controller to handle generating the 'ResponseEntity' object.

Updated the 'GetInfoService' 'getDevicesByLocation' method to return a list of 'GetDeviceResponse' objects rather than a ResponseEntity object.